tests/modules: test more of the timeline/cluster logic
